bug fix in FITS write

+++ @SIM/debias.m
@@ -98,7 +98,7 @@ end
 
 % subtract super bias
 % and combine the mask of the image with that of the bias
-Sim(NotIsBias) = bfun2sim(double(Sim(NotIsBias)),BiasSim,@minus,'MaskFun',@mask_add,'MaskFunPar',{@bitor});
+Sim(NotIsBias) = bfun2sim(Sim(NotIsBias),BiasSim,@minus,'MaskFun',@mask_add,'MaskFunPar',{@bitor});
 
 % subtract bias overscan
 BiasOverScan = [];
